+++ test_files/unit/settings-tab-navigation.test.ts
@@ -28,7 +28,7 @@ describe("SETTINGS_TAB_NAMES", () => {
       "Display",
       "Media",
       "Article saving",
-      "Filters",
+      "Rules",
       "Highlights",
       "Import/Export",
       "Tags",